pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "quicksync"
version = "0.1.0"
description = "A tool that syncs contacts from MockAPI to Mailchimp"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "httpx>=0.24.1",
    "pydantic>=2.0.0",
    "pydantic[email]>=2.0.0",
    "mailchimp-marketing>=3.0.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
quicksync = "quicksync.src.run:main"

[project.optional-dependencies]
test = [
    "pytest>=7.3.1",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0"
]

dev = [
    "black>=23.3.0",
    "ruff>=0.0.270"
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.hatch.build.targets.wheel]
packages = ["quicksync/src"]